fix: move FullAppBenchmarkCaseTrait to its own file for PSR-4 autoloading
FullAppBenchmarkCase.php defined both the FullAppBenchmarkCase class and
a same-file FullAppBenchmarkCaseTrait. Composer's default (non-optimized)
PSR-4 autoloader only resolves a symbol whose name matches its containing
file, so the trait was unresolvable at runtime unless the containing file
happened to be required first — this repo doesn't set
optimize-autoloader, and CI's composer update doesn't pass -o, so
PHPBench's reflector process (which autoloads the benchmark classes
fresh) fataled with "Trait ... not found" for every PHPBench-only
subclass (AsynchronousStackBenchmark, HttpStackBenchmark,
EventSourcingBenchmark, BootingEcotoneBenchmark, KernelBootBenchmark).

Move the trait into its own FullAppBenchmarkCaseTrait.php, matching this
codebase's one-symbol-per-file convention everywhere else.
